large refactor and abstraction
