+++ graphql/resolvers/booking.js
@@ -1,0 +1,37 @@
+const Event = require("../../models/event")
+const Booking = require("../../models/booking")
+
+const { transformBooking, transformEvent } = require("./merge")
+
+module.exports = {
+  bookings: async () => {
+    try {
+      const bookings = await Booking.find()
+      return bookings.map((booking) => {
+        return transformBooking(booking)
+      })
+    } catch (err) {
+      throw err
+    }
+  },
+
+  bookEvent: async (args) => {
+    const fetchedEvent = await Event.findOne({ _id: args.eventId })
+    const booking = new Booking({
+      user: "62ad36872f5a4044773b8026",
+      event: fetchedEvent,
+    })
+    const result = await booking.save()
+    return transformBooking(result)
+  },
+  cancelBooking: async (args) => {
+    try {
+      const booking = await Booking.findById(args.bookingId).populate("event")
+      const event = transformEvent(booking._doc.event)
+      await Booking.deleteOne({ _id: args.bookingId })
+      return event
+    } catch (err) {
+      throw err
+    }
+  },
+}
