+++ graphql/resolvers/merge.js
@@ -1,0 +1,70 @@
+// merging data
+const Event = require("../../models/event")
+const User = require("../../models/user")
+const { dateToString } = require("../../helpers/date")
+
+// name 'Event' anything
+// _id is a mongoDB id, ID is type
+// '!' means it is required/not nullable
+// this just cleans up what you're sending RootMutation.createEvent params
+// return an array of events !needs to be non-nullable/Event
+const events = async (eventIds) => {
+  try {
+    const events = await Event.find({ _id: { $in: eventIds } })
+    return events.map((event) => {
+      return transformEvent(event)
+    })
+  } catch (err) {
+    throw err
+  }
+}
+
+const singleEvent = async (eventId) => {
+  try {
+    const event = await Event.findById(eventId)
+    return transformEvent(event)
+  } catch (err) {
+    throw err
+  }
+}
+
+// fetch a user by its id
+const user = async (userId) => {
+  try {
+    const user = await User.findById(userId)
+    return {
+      ...user._doc,
+      _id: user.id,
+      createdEvents: events.bind(this, user._doc.createdEvents),
+    }
+  } catch (err) {
+    throw err
+  }
+}
+
+const transformEvent = (event) => {
+  return {
+    ...event._doc,
+    _id: event.id,
+    date: dateToString(event._doc.date),
+    creator: user.bind(this, event.creator),
+  }
+}
+
+const transformBooking = (booking) => {
+  return {
+    ...booking._doc,
+    _id: booking.id,
+    user: user.bind(this, booking._doc.user),
+    event: singleEvent.bind(this, booking._doc.event),
+    createdAt: dateToString(booking._doc.createdAt),
+    updatedAt: dateToString(booking._doc.updatedAt),
+  }
+}
+
+exports.transformEvent = transformEvent
+exports.transformBooking = transformBooking
+
+// exports.user = user
+// exports.events = events
+// exports.singleEvent = singleEvent
